new policy regarding assert
assert should only assert something that is explicitly controllable
by program. For example, an input matrix that requires channel == 1
should be asserted (programmer should be responsible to get it right).
Others that may be occurred by system error should not be asserted.
For example, whether or not fd is 0 is determined by system (e.g.
not-found error, no-permission error all can return 0) but previously
in ccv_read, it is asserted. That has been fixed in the commit.