a "wow" error in ccv_dense_matrix_renew
ccv_dense_matrix_renew is an old function. It meant to check whether
the pass-in matrix is on specification, if it is, sign it and return,
if it is not, create/retrieve a new one based on the signature and
return. However, I made a huge mistake by not check if signature is
empty or not, and by the signature generation process, it will generate
a non-empty signature for pass-in empty ones, and using this one
to retrieve matrix from cache. Since I don't use that much
empty-signature matrix myself, the problem hasn't surfaced until
when I have to regularly ccv_make_matrix_mutable/immutable for
dpm implementations.