a bug in varadic parameter of ccv_cache_generate_signature
ccv_cache_generate_signature recognizes the end of the signature by the
ending 0. However, a 0 is 32-bit integer, and we use type uint64_t for
signature, that will cause a problem especially on 32-bit system. This
diff fixed that problem by pass in a typed 0.

no type safe varadic parameter is just evil, lucky that ccv_read is type
safe.