Make sure that we can drive the graph with intermedium nodes as terminal nodes.
In this way, it is easier to make sure intermedium nodes generated
results can be preserved. More over, it will be useful to drive the
graph if there are multiple gradients that we are interested in (because
.we don't know which nodes are actually terminal in this case, we can
only grab nodes based on the gradients we are interested in from
auto-differentiation).