Enable co-allocating 3 symbols
Previously, the allocation algo works with 2 symbol (one symbol and a
companion). It insert these two into the allocation graph and that is
that.

This commit enabled allocation for 3 symbols, so that we can allocate a
companion_ref, a found bigger size companion, and itself. In this way, a
normal conv1 -> smaller conv2 -> bigger3 -> smaller4 -> conv5 (ref
conv1) loop can be allocated more efficiently (with previous algo, it
allocates conv1 and conv5 together, and then smaller2 and smaller4
together, and then bigger3. With co-allocating 3 symbols at the same
time, bigger3, conv1 and conv5 will be taken out first, and smaller2,
smaller4 will be allocated to the same region too.

In above example, if conv5 is not a reference to conv1 (in a loop), then
our original algorithm will perform correctly:

bigger3 and conv1 will first be allocated to the same region, then
conv5 will pair up with smaller2 or just itself. Because itself can be
allocated without new allocation, it will be allocated first (to conv1,
bigger3 region), and then smaller2, smaller4 will be allocated to the
same region.