Fix visiting node that doesn't end with destinations
Previously, the graph visit could end up with nodes that doesn't end
with destinations due to flood fill. This will cause issues if we do
partial run or partially compile a symbolic graph as shown in
some cnnp.core.tests. Fixed with two more passes over the graph. topsort
is hard and I really need to cache the order if ever I want to have this
done faster than it is right now.