Rename cnnp model functions to take name parameter
Each cnnp model (or layer) will take a name. Later this will help
persistence such that a new model can load data from an old model and
the system can figure out which trainable / retainable corresponds to
which layer / model. This will help in case we need to upgrade the model
for training.