Fix the memory corruption issue with imdb.c example
The memory corruption can only be reliably reproduced by commenting out
the tensor_arena_reinit inside ccv_cnnp_model_absorb function.
cuda-memcheck doesn't report error properly inside cublas which
complicates the debug.

The reason is because the different order of minimizer applied from the
original model (applied after we set minimizer) and the new model
(applied when we call gradient init). That will results different tensor
orders. Since we check the match of tensors in the previous absorb
function, that means for some it will be wrong. Make matter worse, we do
the check for inputs and output separately, that means for some exec,
the input matches and updated but the output doesn't.

That scenario will be problematic for GEMM_BACKWARD because we are a bit
loose in checking data, and didn't assert the h_rows equal to g_rows. If
they don't match, we will use h_rows, and in turn, if h_rows is larger,
cause cublas writes memory out of bounds.

The fix is pretty simple, we assert equality of input size / output
size, and just blindly update them without checking the tensor symbol
matches.