Distinguish requirements for not tensor view and not contiguous
Some of the ops requires underlying tensor to be contiguous. Previously,
we check this by checking if it is a tensor view. However, a tensor view
can be contiguous under certain conditions (for example, any slice of a
vector).

This PR makes the first effort to distinguish between these two.

It is a continuous effort to make sure reshaped for tensor variable
would work as expected.