Don't allow create alias if the tensor itself is not fully-specified.
This is in preparation to allow alias an alias if it is contiguous. I
plan to lift the restriction on cannot alias an alias, but with a twist:
you can only alias another alias if the other alias is contiguous. In
this way, I believe the new alias is still fully specified, and can
unlock the cases where we do [1, 0, ...].reshaped(.C(10)).