Implement ccv_nnc_tensor_variable_detach.
This function can be useful if we want to keep the computed variables
for extended time, such as rather than copied, we keep them and later
use them for training purposes. There are effectively two ways to
achieve this, either carefully wrap around with no grad (thus, make sure
the resulting tensor has no symbol associated). Or use this new
function, which seems to be easier to understand (you just "detach" it).