Remove the logic when disjoin, connect to outgoing for destination
Previously, this logic complicates how we actually use it. If we want to
do that, we can use the public API to get the outgoing nodes and concat
them.

Checked the code where we use disjoin, we never made that assumption
(that the source will connect with dest's outgoing nodes).