Use a shared queue for stream context
This implements simple logic to use a shared queue for stream context.
Because each op are scheduled in order, this should be enough to
guarantee correctness without using signal / event.

If there is a need, I can mupltiplex stream contexts onto multiple
queues and use signals / events to synchronize between.