Fix memory synchronization between MPS and MPSGraph
Ideally, I should be able to use MTLFence to do the same. However, it is
not exposed from either MPS and MPSGraph. This hack allows us to encode
the MPSGraph access and MPS on the same command buffer, therefore, make
sure the underlying MPSGraph would synchronize properly for the memory
access.

This hack now enabled us to use SD v2.1 768-v with upcast fix and keep
memory usage down as well.